Add the "eclipse-helper" mojo
The idea of the EclipseHelper in scijava-common is to help develop SciJava
projects in Eclipse, despite the fact that annotation processors -- the
technique on which SciJava plugins are founded, and which makes them so
convenient -- are blatantly ignored by Eclipse.

When importing Maven projects into Eclipse, we have the option, however,
to enforce executing specific Maven goals at specific times in the life
cycle. In our case, we want to force SciJava annotation parsing at the
'process-classes' phase. The mojo introduced in this commit provides the
functionality for this.

To prevent problems e.g. when compiling scijava-common itself, we only
kick into action after verifying that scijava-common is actually a
dependency.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>