Add a class to read and write pom.xml files
This will come in real handy for the upcoming "bump" mojo to bump versions
in POMs. The job of mojo was previously performed by SciJava scripts'
bump-version-in-pom.sh script.

The most important part is that we modify the file contents as little as
possible, to avoid hiding the real changes between unsubstantial
formatting changes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>