PomEditor: allow to modify versions of referenced projects
The procedure established in the SciJava project (which serves as a guinea
pig of robust, collaborative development of scientific software) is to
define dependencies' versions as properties and then use those properties
in the dependencyMangement section for maximal convenience of consumers of
those POMs.

As a consequence, the SciJava project needs tooling to adjust the
properties defining dependency versions every once in a while, and the
upcoming 'bump' mojo of the scijava-maven-plugin is intended to fulfill
that purpose.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>