Add groupIds option
The groupIds flag is an inclusion list of one or more groupIds. The
SnapshotFinder will only report errors if they are in a project whose
groupId is contained in this list.

Note that all groupIds will still be explored recursively.

The default value is an empty list, which indicates all groupIds should
be checked.