Add verbose flag
When verbose is true, every path to a failing gav from the base project
will be reported.

This can be an extreme amount of information, and will likely contain a
significant amount of repeated information when a shared dependency is
bad.

When verbose is false, only direct dependencies of the base pom that
failed will be reported, along with the corresponding set of bad
dependencies or parents that caused failure.