Skip version ranges
If a dependency has a version range, such as [x,y), it is a) not
reproducible and b) un-resolvable. Certainly we could make a best guess
at which version to retrieve, but since we know this brings
reproducibility into question we instead simply mark it as an error and
skip that dependency for recursion.