SnapshotFinder: Convert to visitor
The Maven API provides a visitor interface for traversing dependency
trees. This is used, for example, by the dependency:tree target.

As our custom, fully-recursive mechanism of traversal was prohibitively
slow, we have refactored to a new extensible visitor implementation that
runs an order of magnitude faster.

Note that there may be slight discrepencies in reported errors, as
dependencies are now resolved into a tree by maven instead of manually
searching dependency hierarchies.