SnapshotFinder: move parent pom check
We now call into checkParent multiple times in SnapshotFinder. We had
one of these calls guarded with an if pom.hasParent check, but not the
other - resulting in the potential for an NPE if the base project had no
parent.

It makes much more sense to move the null parent guard into checkParent
itself, which simplifies the code and fixes the NPE.