Prefer conflict-resolved maven dependencies
The AbstractResolveDependencies class, when building a list of Maven
dependencies to check, will now use the "getRelatedArtifact" method if
it is available. This will return the coalesced version of an artifact
when there is a conflict between dependencies with the same groupId and
artifactId, but different versions.

This fixes the BanDuplicateClasses to no longer detect conflicts between
multiple versions of the same GA combination.