Add findEffectiveReactor method
The build reactor determined by Maven is constructed based on execution.
This means that multi-module projects will not be part of the same
reactor if Maven is invoked from an individual submodule.

The findEffectiveReactor method makes a best-effort to discover the
complete intended reactor in a multi-module project by following all
parent and module linkages.