Use effective reactor in SnapshotFinder
The RequireReproducibleBuilds enforcer rule and VerifyNoSnapshots mojo
now attempt to use the effective reactor instead of the invocation-based
reactor as defined by the framework.

This should allow multi-module projects that maintain internal SNAPSHOT
couplings to build individual submodules, provided:
 * They are fully connected via parent/module linkages
 * The complete multi-module project is on disk

This is necessary for use cases such as the Eclipse m2e plugin, which
flattens multi-module projects to individual builds.

Note that the Maven model allows multi-module configuration where the
aggregator is not a parent of any of its modules. In this case, since
the submodules have no reference to their aggregator, SNAPSHOT couplings
must be prohibited.

Closes #12