Add enforcer rule to require elements in the POM
Previously, we noted in the description of our base POMs that inheriting
projects needed to override a boatload of elements. But nothing enforced
it, so everyone was free to ignore that advice and instead unknowingly
inherit the configuration of the parent, which was often incorrect.

This rule makes the requirement to override all of these elements
iron-clad, avoiding this sort of confusion henceforth.