Install dependencies into subdirectories
The mojos use the scijava.app.subdirectory property to determine where
an artifact should be installed. This commit adds the same functionality
for dependencies. If the property is not provided by a dependency, it
falls back to the old case logic. This way, older versions of e.g.
Bio-Formats will be supported as well.