Enumerate native classifiers more exhaustively
Now we generate all combinations of OS family + architectures matching
all observed naming conventions. It's still a hack, but these suffixes
are highly unlikely to be part of any version number.

And we use these generated platform strings in two places: the version
pattern regex, and the default list of subdirectory classifiers.