Simplify encroaching version logic
This change eliminates the general regex in favor of building an
artifact-specific regex, which is more likely to match only file paths
of actually encroaching versions. We also now check only the target
installation directory, as opposed to all directories, for encroaching
versions, which should help eliminate certain types of false positives.