KnownPlatforms: generalize name of utility method
The name nativeClassifierPatterns() suggests the SubdirectoryPattern
list will always only consist of patterns that match particular
classifiers. While this is currently the case, in future it could be
that patterns besides *:*:*:C are possible, so a better name to use is
nativeSubdirectoryPatterns(). This will also avoid confusion with
another method we will need shortly: nativeClassifiers().